Create Assignment5Qno1
Dynamic Programming: Minimum Coin Change Problem
A vending machine needs to return the minimum number of coins as change. The machine should be optimized to return the least number of coins for a given amount.

Task:
Given a list of coin denominations and an amount, implement a function using DP to find the minimum number of coins required to make the amount.