Create Assignment5Qno2
You are building a weather monitoring system that records temperature readings. You need to find the K-th closest reading to a target temperature for accurate prediction adjustments.

Task:
Given a list of sensor readings and a target value, write a program to find the K-th closest element using the Quickselect algorithm.